Iter9: split agent logic
